pyproject.toml
[project]
name = "textbook-analysis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "crawl4ai>=0.8.0",
    "fake-useragent>=2.2.0",
    "google-genai>=1.58.0",
    "marimo>=0.19.2",
    "nbformat>=5.10.4",
    "openai>=2.15.0",
    "pymupdf>=1.26.7",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "ruff>=0.14.11",
    "streamlit>=1.53.0",
]

[dependency-groups]
dev = [
    "ty>=0.0.12",
]
